# Rename `AsciiWordChar` and optimize `Codepoint`