--- a/benches/distributions.rs
+++ b/benches/distributions.rs
@@ -51,7 +51,7 @@ distr!(distr_uniform_i64, i64, Uniform);
 distr!(distr_uniform_i128, i128, Uniform);
 
 distr!(distr_uniform_bool, bool, Uniform);
-distr!(distr_uniform_ascii_char, char, AsciiWordChar);
+distr!(distr_uniform_alphanumeric, char, Alphanumeric);
 
 distr!(distr_uniform01_float32, f32, Uniform01);
 distr!(distr_closed01_float32, f32, Closed01);