--- a/benches/distributions.rs
+++ b/benches/distributions.rs
@@ -52,6 +52,7 @@ distr!(distr_uniform_i128, i128, Uniform);
 
 distr!(distr_uniform_bool, bool, Uniform);
 distr!(distr_uniform_alphanumeric, char, Alphanumeric);
+distr!(distr_uniform_codepoint, char, Codepoint);
 
 distr!(distr_uniform01_float32, f32, Uniform01);
 distr!(distr_closed01_float32, f32, Closed01);